refactor: base component name and folder structure
