--- src/base-component/index.jsx
@@ -1,2 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export { default as BaseComponent } from './BaseComponent';
