+++ src/base-container/components/welcome-page-layout/AuthMediumLayout.jsx
