+++ src/base-container/components/welcome-page-layout/AuthSmallLayout.jsx
