+++ src/forgot-password/ForgotPasswordPage.jsx
@@ -21,7 +21,7 @@ import { forgotPassword, setForgotPasswordFormData } from './data/actions';
 import { forgotPasswordResultSelector } from './data/selectors';
 import ForgotPasswordAlert from './ForgotPasswordAlert';
 import messages from './messages';
-import { BaseComponent } from '../base-component';
+import BaseContainer from '../base-container';
 import { FormGroup } from '../common-components';
 import { DEFAULT_STATE, LOGIN_PAGE, VALID_EMAIL_REGEX } from '../data/constants';
 import { updatePathWithQueryParams, windowScrollTo } from '../data/utils';
@@ -95,7 +95,7 @@ const ForgotPasswordPage = (props) => {
   );
 
   return (
-    <BaseComponent>
+    <BaseContainer>
       <Helmet>
         <title>{formatMessage(messages['forgot.password.page.title'],
           { siteName: getConfig().SITE_NAME })}
@@ -163,7 +163,7 @@ const ForgotPasswordPage = (props) => {
           </Form>
         </div>
       </div>
-    </BaseComponent>
+    </BaseContainer>
   );
 };
